heart_reader/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.2.0
wfdb>=4.1.0
PyYAML>=6.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
onnx>=1.14.0
onnxruntime>=1.15.0
onnxscript>=0.1.0
tensorflow>=2.13.0
tf2onnx>=1.14.0
